fix(onboaring): cleaning boarding flow
